create component to close button and fix shadow under the close button
