+++ ts/navigation/components/CloseButton.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { IOVisualCostants, IconButton } from "@pagopa/io-app-design-system";
+import { useNavigation } from "@react-navigation/native";
+import { View } from "react-native";
+import I18n from "../../i18n";
+
+const CloseButton = () => {
+  const navigation = useNavigation();
+
+  return (
+    <View style={{ marginRight: IOVisualCostants.appMarginDefault }}>
+      <IconButton
+        icon="closeMedium"
+        color="neutral"
+        onPress={() => {
+          navigation.goBack();
+        }}
+        accessibilityLabel={I18n.t("global.buttons.close")}
+      />
+    </View>
+  );
+};
+export default CloseButton;
